feat: adapt online key size
